fix #103: User가 UserDetails를 구현
- CustomUserDetails 삭제
- CustomUserDetailsService에서 User를 리턴하도록 변경
- NotFoundUserException 메시지 생성 방법 수정
