+++ src/main/java/fastcampus/team7/Livable_officener/global/exception/NotFoundUserException.java
@@ -5,11 +5,11 @@ public class NotFoundUserException extends RuntimeException{
     public static final String DEFAULT_MESSAGE = "존재하지 않는 회원입니다.";
 
     public NotFoundUserException() {
-        this(DEFAULT_MESSAGE);
+        super(DEFAULT_MESSAGE);
     }
 
-    public NotFoundUserException(String s) {
-        super(s);
+    public NotFoundUserException(String username) {
+        super(DEFAULT_MESSAGE + " username: " + username);
     }
 
 }
